Refactor DTO and introduce ClerkSniffer component; remove ScriptExecutor and update background script handling
