+++ src/main.tsx
@@ -1,10 +1,10 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import ScriptExecutor  from './ScriptExecutor'
+import ClerkSniffer  from './components/ClerkSniffer'
 import './index.css'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <ScriptExecutor />
+    <ClerkSniffer />
   </StrictMode>
 )
