+++ src/scripts/contentScript.ts
@@ -20,5 +20,6 @@ window.addEventListener('message', (event) => {
     if(event.data.type !== DTO.ClerkSniffer) {
         return;
     }
+    console.log("send", event.data);
     chrome.runtime.sendMessage(event.data);
 });
